Update motionpath.js with offset param
